fix: separate swagger settings from main.ts to a dedicated file
